requirements.txt
pytorch-lightning
torch
torchvision
boto3
gitpython